Extract a method to return the first suitable name field
We made a method to get the first available transliteration of the person
name to one of the languages sent as an argument or the English version
if none is available, and use that as the contents of the name field. We did
this to reduce the complexity of the data method and fix a Rubocop
complexity warning.