Do not overwrite direct link title with name of redirecting link
Sometimes an article is linked to via a redirect link _and_ a direct link.
This commit fixes a problem where the title of a redirected link
overwrites the title of a direct link.

For example, in https://en.wikipedia.org/wiki/Life,_the_Universe_and_Everything,
the article of The Meaning of Liff is linked to by the 'The Meaning of Liff'
link and by the redirect link 'The Deeper Meaning of Liff.' Previously,
wikidata-fetcher was only returning `{'The Deeper Meaning of Liff' => 'Q875382'}`.

This commit returns both links with the wikidata id:
```
{ 'The Meaning of Liff' => 'Q875382', 'The Deeper Meaning of Liff' => 'Q875382' }
```