Form VCR cassette names by joining subject titles
Previously, cassette names were formed from the first title in the subject
list.

This change ensures that cassettes for single titles are not overwritten by
cassettes for mutiple title, where the previous item is first in the list. Now
the title is formed from all the titles in the subject list.